# Conflicts:
#	server/src/main/java/de/materna/dmn/tester/persistence/PersistenceDirectoryManager.java
#	server/src/main/java/de/materna/dmn/tester/servlets/workspace/beans/Workspace.java

+++ server/src/main/java/de/materna/dmn/tester/persistence/PersistenceFileManager.java
@@ -23,9 +23,11 @@ public class PersistenceFileManager {
 	}
 
 	public void removeFile() throws IOException {
-		Files.delete(file);
+		if (fileExists()) {
+			Files.delete(file);
+		}
 	}
-	
+
 	public boolean fileExists() {
 		return Files.exists(file);
 	}
